+++ database/analysis_tables.go
@@ -32,12 +32,12 @@ func (db *DB) createThreatMixtapeTable(ctx context.Context) error {
 			-- **** THREAT INDICATORS ****
 			-- BEACONING
 			beacon_type LowCardinality(String),
-			beacon_score Float32,
-			beacon_threat_score Float32,
-			ts_score Float32,
-			ds_score Float32,
-			dur_score Float32,
-			hist_score Float32,
+			beacon_score Float64,
+			beacon_threat_score Float64,
+			ts_score Float64,
+			ds_score Float64,
+			dur_score Float64,
+			hist_score Float64,
 			ts_intervals Array(Int64),
 			ts_interval_counts Array(Int64),
 			ds_sizes Array(Int64),
@@ -45,41 +45,41 @@ func (db *DB) createThreatMixtapeTable(ctx context.Context) error {
 			
 			-- LONG CONNECTIONS
 			total_duration Float64,
-			long_conn_score Float32,
+			long_conn_score Float64,
 
 			-- STROBE
-			strobe_score Float32,
+			strobe_score Float64,
 
 			-- C2 OVER DNS
 			subdomain_count UInt64,
-			c2_over_dns_score Float32,
-			c2_over_dns_direct_conn_score Float32,
+			c2_over_dns_score Float64,
+			c2_over_dns_direct_conn_score Float64,
 
 			-- THREAT INTEL
 			threat_intel Bool,
-			threat_intel_score Float32,
+			threat_intel_score Float64,
 
 			-- **** MODIFIERS ****
 			modifier_name LowCardinality(String),
-			modifier_score Float32,
+			modifier_score Float64,
 			modifier_value String,
 
 			-- PREVALENCE
 			prevalence_total UInt64,
-			prevalence Float32,
-			prevalence_score Float32,
+			prevalence Float64,
+			prevalence_score Float64,
 			network_size UInt64,
 
 			first_seen_historical DateTime(),
-			first_seen_score Float32,
+			first_seen_score Float64,
 
 			-- THREAT INTEL DATA SIZE
-			threat_intel_data_size_score Float32,
+			threat_intel_data_size_score Float64,
 
 
 			-- MISSING HOST HEADER
 			missing_host_count UInt64,
-			missing_host_header_score Float32
+			missing_host_header_score Float64
 
 		) ENGINE = MergeTree()
 		PRIMARY KEY (analyzed_at, dst_nuid, src_nuid, src, fqdn, dst, hash)
